fields unpacking

+++ Sources/SkyrimFileFormat/Field.swift
@@ -9,13 +9,13 @@ import Foundation
 
 class Field: CustomStringConvertible {
     let type: Tag
-    let size: UInt32
+    let size: UInt16
     let data: Bytes
 
     init<S: AsyncIteratorProtocol>(_ iterator: inout AsyncBufferedIterator<S>) async throws where S.Element == UInt8 {
         let tag = try await iterator.next(littleEndian: UInt32.self)
         self.type = Tag(tag)
-        self.size = try await iterator.next(littleEndian: UInt32.self)
+        self.size = try await iterator.next(littleEndian: UInt16.self)
         self.data = try await iterator.next(bytes: Bytes.self, count: Int(size))
     }
